Reuse pie chart color palettes
