+++ packages/widgets/src/components/PieChart/index.ts
@@ -1,3 +1,4 @@
 export * from './PieChart';
 export * from './PieChartLegend';
+export * as pieChartPalettes from './colorPalettes';
 export * from './types';
